Move get_playlist_repost_intersection_users
