+++ discovery-provider/src/queries/queries.py
@@ -30,6 +30,7 @@ from src.queries.get_feed import get_feed
 from src.queries.get_repost_feed_for_user import get_repost_feed_for_user
 from src.queries.get_follow_intersection_users import get_follow_intersection_users
 from src.queries.get_track_repost_intersection_users import get_track_repost_intersection_users
+from src.queries.get_playlist_repost_intersection_users import get_playlist_repost_intersection_users
 
 
 logger = logging.getLogger(__name__)
@@ -132,45 +133,12 @@ def get_track_repost_intersection_users_route(repost_track_id, follower_user_id)
 # - Followee = user that is followed. Follower = user that follows.
 # - repost_playlist_id = playlist that is reposted. repost_user_id = user that reposted playlist.
 @bp.route("/users/intersection/repost/playlist/<int:repost_playlist_id>/<int:follower_user_id>", methods=("GET",))
-def get_playlist_repost_intersection_users(repost_playlist_id, follower_user_id):
-    users = []
-    db = get_db_read_replica()
-    with db.scoped_session() as session:
-        # ensure playlist_id exists
-        playlist_entry = session.query(Playlist).filter(
-            Playlist.playlist_id == repost_playlist_id,
-            Playlist.is_current == True
-        ).first()
-        if playlist_entry is None:
-            return api_helpers.error_response('Resource not found for provided playlist id', 404)
-
-        query = (
-            session.query(User)
-            .filter(
-                User.is_current == True,
-                User.user_id.in_(
-                    session.query(Repost.user_id)
-                    .filter(
-                        Repost.repost_item_id == repost_playlist_id,
-                        Repost.repost_type != RepostType.track,
-                        Repost.is_current == True,
-                        Repost.is_delete == False
-                    )
-                    .intersect(
-                        session.query(Follow.followee_user_id)
-                        .filter(
-                            Follow.follower_user_id == follower_user_id,
-                            Follow.is_current == True,
-                            Follow.is_delete == False
-                        )
-                    )
-                )
-            )
-        )
-        users = paginate_query(query).all()
-        users = helpers.query_result_to_list(users)
-
-    return api_helpers.success_response(users)
+def get_playlist_repost_intersection_users_route(repost_playlist_id, follower_user_id):
+    try:
+        users = get_playlist_repost_intersection_users(repost_playlist_id, follower_user_id)
+        return api_helpers.success_response(users)
+    except Exception as e:
+        return api_helpers.error_response(str(e), 404)
 
 
 # Get paginated users that follow provided followee_user_id, sorted by their follower count descending.
